feat(api): dev space adaption

+++ pkg/nocalhost-api/app/api/v1/application_user/application_user.go
@@ -1,5 +1,5 @@
 package application_user
 
 type ApplicationUsersRequest struct {
-	Users   []uint64 `json:"users" validate:"required"`
+	Users   []int `json:"users" validate:"required"`
 }
